refactor(dashboard): extract dashboardFactory
